+++ src/constants/index.ts
@@ -1,4 +1,4 @@
-export const ramenyaType = [
+export const RAMENYA_TYPES = [
   "이에케",
   "돈코츠",
   "소유",
@@ -11,3 +11,11 @@ export const ramenyaType = [
   "토리",
   "츠케멘",
 ];
+
+export enum OpenStatus {
+  OPEN = "영업 중",
+  CLOSED = "영업 종료",
+  BREAK = "쉬는 중",
+  BEFORE_OPEN = "영업 전",
+  DAY_OFF = "휴무일",
+}
